appInfo2 enabled capability to retrieve build info from database

+++ src/main/java/com/jhart/repo/buildinfo/BuildInfoRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import com.jhart.domain.BuildInfo;
 
 @Repository("BuildInfoRepository")
-public interface BuildInfoRepository extends JpaRepository<BuildInfo, Long>{
+public interface BuildInfoRepository extends JpaRepository<BuildInfo, Integer>{
 
 }
